+++ product/TTV/Form/index.js
@@ -6,7 +6,7 @@ import * as Yup from 'yup';
 import {AppRegistry} from 'react-native';
 import { useState } from 'react';
 import App from './App';
-import Form from './Validator/Form';
+import TabBar from './Validator/TabNavigation';
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => Form);
+AppRegistry.registerComponent(appName, () => TabBar);
